Input validator restructuring
