+++ common/modules/module_utils/input_validation/common_utils/logical_validation.py
@@ -15,9 +15,9 @@
 import sys
 sys.path.append("module_utils/validation_flows")
 
-from ansible.module_utils.validation_flows import provision_validation
-from ansible.module_utils.validation_flows import common_validation
-from ansible.module_utils.validation_flows import roles_validation
+from ansible.module_utils.input_validation.validation_flows import provision_validation
+from ansible.module_utils.input_validation.validation_flows import common_validation
+from ansible.module_utils.input_validation.validation_flows import roles_validation
 
 # L2 Validation Code - validate anything that could not have been validated with JSON schema
 # Main validation code that calls one of the validation functions based on the tag(s) used. input_file_inventory in validate_input.py contains dict of the tags being called.
