--- common/modules/module_utils/input_validation/credential_rules.json
@@ -1,104 +1,0 @@
-{
-      "provision_password": {
-        "minLength": 8,
-        "maxLength": 32,
-        "pattern": "^[^-,\\'\"]+$",
-        "description": "Password required during OS provisioning for root users. Length must be at least 8 characters and must not contain commas (,), hyphens (-), single quotes ('), or double quotes (\\\")."
-      },
-      "postgresdb_password": {
-        "minLength": 8,
-        "maxLength": 32,
-        "pattern": "^[a-zA-Z0-9]+$",
-        "description": "Password required for PostgreSQL database. Length must be at least 8 characters and must contain only alphanumeric characters (letters and numbers)."
-      },
-      "bmc_username": {
-        "minLength": 1,
-        "maxLength": 64,
-        "pattern": "^[^\\\\\\-'\"]+$",
-        "description": "Username for BMC (idrac) access. The same credentials must be used across all servers."
-      },
-      "bmc_password": {
-        "minLength": 3,
-        "maxLength": 32,
-        "pattern": "^(?!.*[-\\\\'\"]).*$",
-        "description": "Password required for BMC (idrac) access. Length must be at least 3 characters and must not contain commas (,), hyphens (-), single quotes ('), or double quotes (\\\")."
-      },
-      "switch_snmp3_username": {
-        "minLength": 4,
-        "maxLength": 64,
-        "pattern": "^[^\\\\\\-'\"]+$",
-        "description": "Non-admin SNMPv3 credentials for the PXE switch. If multiple switches are provided, use the same credentials for all switches."
-      },
-      "switch_snmp3_password": {
-        "minLength": 4,
-        "maxLength": 64,
-        "pattern": "^[^-,\\'\"]*$",
-        "description": "Password required for SNMPv3 access to the PXE switch. Must not contain commas (,), hyphens (-), single quotes ('), or double quotes (\\\")."
-      },
-      "pulp_password": {
-        "minLength": 8,
-        "maxLength": 32,
-        "pattern": "^[^-,\\'\"]+$",
-        "description": "Pulp required for setting up Pulp container. Length must be at least 8 characters and must not contain commas (,), hyphens (-), single quotes ('), or double quotes (\\\")."
-      },
-      "docker_username": {
-        "minLength": 4,
-        "maxLength": 32,
-        "pattern": "^[a-z0-9_.]+$",
-        "description": "Username for Dockerhub account. This will be used for Docker login."
-      },
-      "docker_password": {
-        "minLength": 8,
-        "maxLength": 128,
-        "pattern": "^[a-zA-Z0-9!@#$%^&*()_+=,.?<>;:{}\\[\\]|-]{6,128}$",
-        "description": "Password for Dockerhub account. Must be at least 8 characters long and can contain letters, numbers, and special characters."
-      },
-      "slurm_db_password": {
-        "description": "SlurmDB password must not contain special characters like hyphens (-), single quotes ('), double quotes (\") or backslashes (\\).",
-        "pattern": "^(?:[^\\-\\'\\\"\\\\]*)?$",
-        "minLength": 8,
-        "maxLength": 32
-      },
-      "directory_manager_password": {
-        "description": "Password for LDAP Directory Manager. Must not contain hyphens (-), single quotes ('), double quotes (\"), at symbols (@), or backslashes (\\).",
-        "minLength": 8,
-        "maxLength": 32,
-        "pattern": "^[^\\-\\'\\\"@\\\\]*$"
-      },
-      "kerberos_admin_password": {
-        "description": "Password for Kerberos admin in FreeIPA. Must not contain hyphens (-), single quotes ('), double quotes (\"), at symbols (@), or backslashes (\\).",
-        "minLength": 8,
-        "maxLength": 32,
-        "pattern": "^[^\\-\\'\\\"@\\\\]*$"
-      },
-      "openldap_db_username": {
-        "minLength": 4,
-        "maxLength": 64,
-        "description": "Username for OpenLDAP database admin. Must not contain semicolons (;), square brackets ([]), or backticks (`).",
-        "pattern": "^[^;\\[\\]`]+$"
-      },
-      "openldap_db_password": {
-        "description": "Password for OpenLDAP database admin. Must not contain hyphens (-), single quotes ('), double quotes (\"), at symbols (@), or backslashes (\\).",
-        "minLength": 8,
-        "maxLength": 32,
-        "pattern": "^[^\\-\\'\\\"@\\\\]*$"
-      },
-      "openldap_config_username": {
-        "minLength": 4,
-        "maxLength": 64,
-        "description": "Username for OpenLDAP configuration admin. Must not contain semicolons (;), square brackets ([]), or backticks (`).",
-        "pattern": "^[^;\\[\\]`]+$"
-      },
-      "openldap_config_password": {
-        "description": "Password for OpenLDAP configuration admin. Must not contain hyphens (-), single quotes ('), double quotes (\"), at symbols (@), or backslashes (\\).",
-        "minLength": 8,
-        "maxLength": 32,
-        "pattern": "^[^\\-\\'\\\"@\\\\]*$"
-      },
-      "openldap_monitor_password": {
-        "description": "Password for OpenLDAP monitor admin. Must not contain hyphens (-), single quotes ('), double quotes (\"), at symbols (@), or backslashes (\\).",
-        "minLength": 8,
-        "maxLength": 32,
-        "pattern": "^[^\\-\\'\\\"@\\\\]*$"
-      }
-  }
