+++ common/modules/module_utils/input_validation/schema/credential_rules.json
