+++ common/modules/module_utils/input_validation/schema/high_availability_config.json
