+++ common/modules/module_utils/input_validation/schema/network_config.json
