+++ common/modules/module_utils/input_validation/schema/omnia_config.json
