+++ common/modules/module_utils/input_validation/schema/passwordless_ssh_config.json
