+++ common/modules/module_utils/input_validation/schema/provision_config_credentials.json
