+++ common/modules/module_utils/input_validation/schema/roce_plugin_config.json
