+++ common/modules/module_utils/input_validation/schema/roles_config.json
