+++ common/modules/module_utils/input_validation/schema/security_config.json
