+++ common/modules/module_utils/input_validation/schema/software_config.json
