+++ common/modules/module_utils/input_validation/schema/storage_config.json
