+++ common/modules/module_utils/input_validation/schema/telemetry_config.json
