--- input_validation/module_utils/schema/network_spec.json
@@ -1,95 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "type": "object",
-  "required": ["Networks"],
-  "properties": {
-    "Networks": {
-      "type": "array",
-      "minItems": 1,
-      "items": {
-        "type": "object",
-        "anyOf": [
-          {
-            "type": "object",
-            "required": ["admin_network"],
-            "properties": {
-              "admin_network": {
-                "type": "object",
-                "required": [
-                  "oim_nic_name",
-                  "netmask_bits",
-                  "primary_oim_admin_ip",
-                  "static_range",
-                  "dynamic_range",
-                  "correlation_to_admin",
-                  "admin_uncorrelated_node_start_ip",
-                  "MTU"
-                ],
-                "properties": {
-                  "oim_nic_name": { "type": "string" },
-                  "netmask_bits": { "type": "string", "pattern": "^(1[0-9]|2[0-9]|[1-9])$|^3[0-2]$" },
-                  "primary_oim_admin_ip": { "type": "string", "pattern": "^\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}$" },
-                  "static_range": { "type": "string", "pattern": "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)-(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$" },
-                  "dynamic_range": { "type": "string", "pattern": "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)-(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$" },
-                  "correlation_to_admin": { "type": "boolean" },
-                  "admin_uncorrelated_node_start_ip": { "type": "string", "pattern": "^\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}$" },
-                  "network_gateway": { "type": "string" },
-                  "DNS": { "type": "string" },
-                  "MTU": { "type": "string" }
-                }
-              }
-            }
-          },
-          {
-          "type": "object",
-          "required": ["bmc_network"],
-          "properties": {
-            "bmc_network": {
-              "type": "object",
-              "properties": {
-                "oim_nic_name": { "type": "string" },
-                "netmask_bits": { "type": "string" },
-                "dynamic_range": { "type": "string" },
-                "dynamic_conversion_static_range": { "type": "string" },
-                "reassignment_to_static": { "type": "boolean" },
-                "network_gateway": { "type": "string" },
-                "MTU": { "type": "string" },
-                "additionalProperties": false
-              },
-              "allOf": [
-                {
-                  "if": {
-                    "properties": {
-                      "oim_nic_name": { "type": "string", "minLength": 1 }
-                    }
-                  },
-                  "then": {
-                    "required": ["netmask_bits", "dynamic_range", "dynamic_conversion_static_range"],
-                    "properties": {
-                      "netmask_bits": {
-                        "type": "string",
-                        "minLength": 1,
-                        "pattern": "^(1[0-9]|2[0-9]|[1-9])$|^3[0-2]$"
-                      },
-                      "dynamic_range": {
-                        "type": "string",
-                        "minLength": 1,
-                        "pattern": "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)-(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
-                      },
-                      "dynamic_conversion_static_range": {
-                        "type": "string",
-                        "minLength": 1,
-                        "pattern": "^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)-(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$"
-                      }
-                    }
-                  }
-                }
-              ]
-            }
-          }
-        }
-        ]
-      }
-    }
-  }
-}
