--- input_validation/module_utils/schema/server_spec.json
@@ -1,64 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "type": "object",
-  "properties": {
-    "Categories": {
-        "type": ["array", "null"],
-        "default": [],
-        "items": {
-            "type": "object",
-            "patternProperties": {
-                "^[a-zA-Z0-9_-]+$": {
-                    "type": "array",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "network": {
-                                "type": "array",
-                                "items": {
-                                    "type": "object",
-                                    "patternProperties": {
-                                        "^[a-zA-Z0-9_-]+$": {
-                                            "type": "object",
-                                            "properties": {
-                                                "nicnetwork": { "type": "string" },
-                                                "nictypes": { "type": "string" },
-                                                "nicdevices": { "type": "string" },
-                                                "metric": { "type": "string" }
-                                            },
-                                            "required": ["nicnetwork", "nictypes"],
-                                            "additionalProperties": false
-                                        }
-                                    }
-                                }
-                            },
-                            "os": {
-                                "type": "array",
-                                "items": {
-                                    "type": "object",
-                                    "properties": {
-                                        "kernel": {
-                                            "type": "array",
-                                            "items": {
-                                                "type": "object",
-                                                "properties": {
-                                                    "cmdline": { "type": "string" }
-                                                },
-                                                "required": ["cmdline"],
-                                                "additionalProperties": false
-                                            }
-                                        }
-                                    }
-                                },
-                                "required": ["kernel"],
-                                "additionalProperties": false
-                            }
-                        },
-                        "additionalProperties": false
-                    }
-                }
-            }
-        }
-    }
-  }
-}
